# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Online Supermarket System API",
    "description": "RESTful API specification for the Multi-branch Online Supermarket System (.NET 10 Minimal APIs).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5072",
      "description": "Local Development Server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Docker Container API Gateway"
    }
  ],
  "paths": {
...
    "/api/users/me/addresses": {
      "get": {
        "tags": ["Addresses"],
        "summary": "Get user delivery addresses",
        "operationId": "GetAddresses",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of user addresses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AddressDto" }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "tags": ["Addresses"],
        "summary": "Create delivery address",
        "operationId": "CreateAddress",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAddressRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Address created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AddressDto" }
              }
            }
          },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/users/me/addresses/{id}": {
      "put": {
        "tags": ["Addresses"],
        "summary": "Update delivery address",
        "operationId": "UpdateAddress",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAddressRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Address updated" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Address not found" }
        }
      },
      "delete": {
        "tags": ["Addresses"],
        "summary": "Delete delivery address",
        "operationId": "DeleteAddress",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Address deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Address not found" }
        }
      }
    },
    "/api/users/me/addresses/{id}/default": {
      "put": {
        "tags": ["Addresses"],
        "summary": "Set address as default",
        "operationId": "SetAddressAsDefault",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Default address set" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Address not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateAddressRequest": {
        "type": "object",
        "required": ["recipientName", "phone", "street", "ward", "district", "city"],
        "properties": {
          "recipientName": { "type": "string" },
          "phone": { "type": "string" },
          "street": { "type": "string" },
          "ward": { "type": "string" },
          "district": { "type": "string" },
          "city": { "type": "string" },
          "postalCode": { "type": "string", "nullable": true }
        }
      },
      "UpdateAddressRequest": {
        "type": "object",
        "required": ["recipientName", "phone", "street", "ward", "district", "city"],
        "properties": {
          "recipientName": { "type": "string" },
          "phone": { "type": "string" },
          "street": { "type": "string" },
          "ward": { "type": "string" },
          "district": { "type": "string" },
          "city": { "type": "string" },
          "postalCode": { "type": "string", "nullable": true }
        }
      },
      "AddressDto": {
        "type": "object",
        "required": ["id", "recipientName", "phone", "street", "ward", "district", "city", "isDefault", "createdAtUtc", "updatedAtUtc"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "recipientName": { "type": "string" },
          "phone": { "type": "string" },
          "street": { "type": "string" },
          "ward": { "type": "string" },
          "district": { "type": "string" },
          "city": { "type": "string" },
          "postalCode": { "type": "string", "nullable": true },
          "isDefault": { "type": "boolean" },
          "createdAtUtc": { "type": "string", "format": "date-time" },
          "updatedAtUtc": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}